backend/requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-multipart==0.0.20          # form data / file uploads

# ─── Configuration ────────────────────────────────────────────────────────────
pydantic-settings==2.7.0
pydantic[email]==2.10.4

# ─── Database ─────────────────────────────────────────────────────────────────
supabase==2.11.0                  # Supabase Python client (sync + async)
asyncpg==0.30.0                   # Direct Postgres for complex queries

# ─── Auth & Security ──────────────────────────────────────────────────────────
PyJWT==2.10.1                     # JWT decode/validation
cryptography==44.0.0              # AES-256-GCM for payment credentials

# ─── HTTP Client ──────────────────────────────────────────────────────────────
httpx==0.28.1                     # Async HTTP (WhatsApp Graph API, webhooks)

# ─── Background Tasks ─────────────────────────────────────────────────────────
celery[redis]==5.4.0
redis==5.2.1
flower==2.0.1                     # Celery monitoring UI

# ─── Rate Limiting ────────────────────────────────────────────────────────────
slowapi==0.1.9

# ─── AI ───────────────────────────────────────────────────────────────────────
anthropic==0.43.0

# ─── Payments ─────────────────────────────────────────────────────────────────
stripe==11.4.1

# ─── Email ────────────────────────────────────────────────────────────────────
resend==2.26.0

# ─── PDF Generation ───────────────────────────────────────────────────────────
reportlab==4.2.5

# ─── Dev / Linting / Testing ──────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
ruff==0.8.4
mypy==1.14.0
